automaton.hpp: Try to make unknown affect position consistent
In Aho-Corasick automaton we know the affect once we traversed the sequence.
However for backward-compatibility reasons with k-mer indexes we need to move
those affects so that they correspond to the start of the sequence.
However for an Unknown Affect, it is hard to tell where we should put it
as the affect is unknown. If we see that we will overwrite an existing
affect, we try another position.

This solves some cases in #4225