backup/Dockerfile: crontab wants the files to belong to root
The file /etc/backup/backup-cron doesn't belong to root so cron didn't want to
run it (even with a symbolic link). We circumvent the issue by regularly
copying the file. Thus we can still modify the file on the host and have the
changes taken into account in the container without having to rebuild it.