Store a different affect for each germline.
See #3584.

This prevents issue with TRA+D and TRD (for instance) where a same file is used in several places.
However the drawback is that we will often have several affectations for a given k-mer (especially for J).

This is a work in progress, there are things to clean/remove in the commit