bioreader: Increase number of sequences sampled to get the approximation
In a compressed stream the size of the buffer is something like 8k or 16k. Thus when calling tellg(), the position we get is a multiple of this figure.
This is an approximation of the actual position. Thus the less sequences we sample, the worst the approximation. That's why we increase 10-fold the number of sequences sampled.