segment.hpp: prevent affect overwriting
When we had two identical affects with a different strand, one could overwrite the other, leading
to an undertermined germline while we should have a normal one. This is now fixed by storing
the strands separately.

On top of that, the strand information is also stored separately in order to test efficiently for a common strand.