algo: Ignore upstream/downstream intronic regions during the deisgnation step
In order to do so, we introduce a flag to determine whether the uppercase letters should be ignored.
This flag is passed to the germlines and then to the BioReader. The germlines take it into account
after the indexing step. Indexing is used for the V(D)J detection step, thus the BioReader are not
used anymore for the detection step but only for the designation step.

We also add an option in order to deactivate the behaviour